requirements.txt
bentoml>=1.0.0
torch>=1.12.0
torchvision>=0.13.0
pillow>=9.0.0
numpy>=1.21.0
joblib
tqdm
pip-chill
wincertstore
